refactor related tests into their own modules
